Implement widget hidden/invisible and bg-pattern change
`hidden` is supposed to be used by user to mark persistent status.
Methods: `set-hidden`, `hide`, `show`.

`invisible` is a temporary status supposed to reflect current widget
visiblity on its parent. For now the only reason for a widget to be
invisible is to fall out of parent's rectange. Not really recommended to
be set by a user unless he really knows what he's doing.