Re-consider child canvas aproach again
Reduce extra redraws as much as possible. Instead, widget's canvas
becomes kind of a cached state after redraw but without child canvas
imprinted. When canvas are flattenned down, the cache is getting cloned,
child canvas with their attached invalidations are imprinted and the
result is then sent over to the parent.

The approach should also reduce the amount of invalidations being passed
around.

Widgets are now coercing to `False` if closed. This simplifies cases
when a child is still accessible on its parent but can't be used. Then
it is sufficient to `if $parent<WidgetName>`.