Optimize the numbers of canvas flattening
Do so similarly to preventing extra redraws: block flattening until a
widget is ready to proceed. In particular, a window would block until
both client and border canvas redrawn when window redraw initiated.
Desktop would block until the screen is ready to print again.

For better responsiveness made Event::Screen::Geom an immediate event.

The window example script was optimized to do the next step only when
window update event is received but not before the previous step is
completed. A few trick were added like closing the static window in the
middle, event reporting in a `TextScroll` widget, some coloring. Border
of the static window is intentionally left transparent, it's not a bug!