Speed up PChannel and make it almost independent of number of priorties
Heavy benchmarks show that the channel is now only 1.5-1.8 times slower
on receiving than the standard Channel even on 1000 priority queues.
Sending is 3-4 times slower but it's ok as long as it's still 2-3 times
faster than receiving. I.e. receive will never wait for a send to
complete in a busy environment.

PChannel has been made literally lock-less. The only lock used is for
pre-allocation of new priority queues. But this operation should
normally happen only once in a channel lifetime if pre-allocation is
used. And even without pre-allocaion it is rather rare op which doesn't
really affect the overall performance.